+++ studio/components/to-be-cleaned/ProductEmptyState.tsx
@@ -37,7 +37,7 @@ const ProductEmptyState: FC<Props> = ({
                   {ctaButtonLabel}
                 </Button>
               </Tooltip.Trigger>
-              {disabled && disabledMessage && (
+              {disabled && disabledMessage.length > 0 && (
                 <Tooltip.Content side="bottom">
                   <Tooltip.Arrow className="radix-tooltip-arrow" />
                   <div
